requirements.txt
# TerrainDreamer Dependencies
# ============================

# Core
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0

# Deep learning (CUDA 12.1)
# pip install torch==2.3.1 --index-url https://download.pytorch.org/whl/cu121
torch==2.3.1

# Simulator
pybullet>=3.2.6            # Physics engine (PyBullet Moon UGV simulator)
pygame>=2.6.0              # Merged single-window UI (3D + LiDAR + map + IMU)
gymnasium>=0.29.0          # RL environment interface (LunarJackalEnv)
stable-baselines3>=2.3.0   # PPO policy training (optional)

# Real hardware interfaces (only needed for Jackal deployment)
pyserial>=3.5              # VectorNav VN-100 IMU serial
python-can>=4.0.0          # CAN bus (vehicle control output)

# Training utilities
tqdm>=4.65.0
tensorboard>=2.12.0
